Move alert types to web app
