+++ unified-codes/apps/web/src/components/containers/index.ts
@@ -1,4 +1,5 @@
 export * from './app';
+export * from './alert';
 export * from './detail';
 export * from './error';
 export * from './explorer';
