+++ unified-codes/libs/data/src/lib/types/index.ts
@@ -1,4 +1,3 @@
-export * from './Alert';
 export * from './ApolloService';
 export * from './AuthenticationService';
 export * from './AuthorisationService';
