+++ unified-codes/libs/ui/src/lib/components/feedback/index.ts
@@ -1,2 +1,1 @@
 export * from './atoms';
-export * from './molecules';
