+++ unified-codes/libs/ui/src/lib/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useAlert } from './useAlert';
 export { useToggle } from './useToggle';
